zmqcheck reimplemented to improve portability
Current `zmqcheck doesn't work on Windows because of multiple issues. One example is using hardcoded `make` instead of `$Config{make}` which can be either `dmake` or `nmake` on Windows.

Fortunately, a more portable and robust implementation is already present in `inc/Devel/CheckLib.pm`.

If this patch is accepted, whole `tools/zmqcheck` folder can be deleted and manifest updated.