ipc:// is not supported on Windows
https://zeromq.jira.com/browse/LIBZMQ-153 - here's official bug report.

I used a rather dirty way of getting a free port instead of `TCP::Test::empty_port()` as I didn't want to introduce a dependency on `TCP::Test` if it's not really needed. And I didn't want to make code much more complicated either.

So now we have a clear improvement on Windows (it kind of works instead of total failure), and on Unix the code didn't change.